# Signalfx exporter: reduce memory allocations for big batches processing