--- a/exporter/signalfxexporter/translation/converter_test.go
+++ b/exporter/signalfxexporter/translation/converter_test.go
@@ -242,7 +242,7 @@ func Test_MetricDataToSignalFxV2(t *testing.T) {
 	c := NewMetricsConverter(logger, nil)
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			gotSfxDataPoints, gotNumDroppedTimeSeries := c.MetricDataToSignalFxV2(tt.metricsDataFn())
+			gotSfxDataPoints, gotNumDroppedTimeSeries := c.MetricDataToSignalFxV2(tt.metricsDataFn(), nil)
 			assert.Equal(t, tt.wantNumDroppedTimeseries, gotNumDroppedTimeSeries)
 			// Sort SFx dimensions since they are built from maps and the order
 			// of those is not deterministic.
@@ -304,7 +304,7 @@ func TestMetricDataToSignalFxV2WithTranslation(t *testing.T) {
 		},
 	}
 	c := NewMetricsConverter(zap.NewNop(), translator)
-	got, dropped := c.MetricDataToSignalFxV2(md)
+	got, dropped := c.MetricDataToSignalFxV2(md, nil)
 
 	assert.EqualValues(t, 0, dropped)
 	assert.EqualValues(t, expected, got)
@@ -462,7 +462,7 @@ func Test_InvalidPoint_NoValue(t *testing.T) {
 		},
 	}
 	c := NewMetricsConverter(logger, nil)
-	_, gotNumDroppedTimeSeries := c.MetricDataToSignalFxV2(metricData)
+	_, gotNumDroppedTimeSeries := c.MetricDataToSignalFxV2(metricData, nil)
 	assert.Equal(t, 1, gotNumDroppedTimeSeries)
 }
 
@@ -480,7 +480,7 @@ func Test_InvalidMetric(t *testing.T) {
 		},
 	}
 	c := NewMetricsConverter(logger, nil)
-	_, gotNumDroppedTimeSeries := c.MetricDataToSignalFxV2(metricData)
+	_, gotNumDroppedTimeSeries := c.MetricDataToSignalFxV2(metricData, nil)
 	// Only 1 timeseries is dropped because the nil metric does not have any timeseries.
 	assert.Equal(t, 1, gotNumDroppedTimeSeries)
 }