**Description:**
Enabling batching causes many extra allocations in signalfx exporter. This change reduces allocations of []*sfxpb.DataPoint slices by pre-initializing []*sfxpb.DataPoint slice for the whole batch.

**Testing:**
Benchmark test on a 1000 metrics batch:

BEFORE:
```
Running tool: /usr/local/bin/go test -benchmem -run=^$ -bench ^(BenchmarkExporterConsumeData)$
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/signalfxexporter
BenchmarkExporterConsumeData-12           78      14264799 ns/op     6911068 B/op      70208 allocs/op
PASS
ok      github.com/open-telemetry/opentelemetry-collector-contrib/exporter/signalfxexporter 1.896s
```

AFTER:
```
Running tool: /usr/local/bin/go test -benchmem -run=^$ -bench ^(BenchmarkExporterConsumeData)$
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/signalfxexporter
BenchmarkExporterConsumeData-12    	      93	  12466935 ns/op	 6429209 B/op	   68854 allocs/op
PASS
ok  	github.com/open-telemetry/opentelemetry-collector-contrib/exporter/signalfxexporter	1.897s
```